requirements.txt
modbus-tk>=1.1.5
pyserial>=3.5
# Dev/test
pytest>=9.0.2
pytest-asyncio>=1.3.0
